Integrated social feature work by Ryan Waggoner.
I tried merging in from https://github.com/ryanwaggoner/INaturalistIOS
but there were insane conflicts with the storyboard and the data
model, so I used checkout on the ObjC source files and tried to
replicate the rest. As far as I can tell this mostly works in the way
Ryan left it.