Bugfix: inf loop during refresh with unsynced obs field values
Overriding the core data accessors is risky business, but we do it for
associates to try and maintain foreign key integrity. What we were not
doing was checking to see if the associate was there before accessing
its key and assigning it to a foreign key attribute, which resulted in
an attempt to assign a nil value to a foreign key attribute, which
Restkit's caching layer was unhappy with and resulted in some infinite
looping.