drain the sync queue using blocks, not delegates
try to avoid crashes when the sync queue fails to get a new current
object, by processing responses to sync API calls in closures that
contain a reference to the relevant current model that's being sync'd.

for clarity & naming safety, replace the current as NSDictionary with an
NSObject that has properties for the model, the synced count, etc.

replace the syncSelector with a loaderConfigBlock, which was actually
what was happening with the syncSelector anyways. it was configuring the
restkit object loader with a POST object containing the observationPhoto
file data and then depending on the sync queue to handle delegate
callbacks. since we aren't doing the delegate thing anymore, we need to
just execute that loader config explicitly.

still TBD: rewrite the SyncQueueDelegate to be a
SyncQueueNotificationDelegate, which it really is, and remove
dependencies on passing around restkit internals. also, consider
replacing the loaderConfigBlock with a more generic postConfigBlock
that can help the SyncQueue construct the necessary POST body without
necessarily including RestKit details.