s/SyncQueueDelegate/SyncQueueNotificationDelegate/
replace the SyncQueueDelegate protocol with a
SyncQueueNotificationDelegate protocol. ObservationsViewController,
which implements this protocol, now uses SVProgressHUD for
notifications. try to expose less of the SyncQueue internals (restkit
object loader, etc) to the SyncQueueNotificationDelegate, which is
primarily responsible for notifying the user about success/failure of
notifications (along with prompting for login when necessary).