fix observation sharing on iPad
on iPad, share via a popover. furthermore, pop share after 1 second, to
allow the "open in safari ORshare" action sheet to dismiss. action
sheets on iPad seem to be implemeneted as presented VCs, and we can't
present multiple view controllers at once. so wait a second to allow the
action sheet to finish dismissing itself before presenting the share
activity view controller (necessary whether in a popover or not).