be defensive around missing obs photos
this defensiveness sucks, but in the event that we get all the way to sync
of an ObservationPhoto and the ImageStore doesn't have a valid photo
file object, then we need to bail on syncing the ObservationPhoto and
notify the user so they can hopefully do something about it.

next step is to beef up the ImageStore saving process.