don't dismiss keyboard on end displaying cell
since we now have done/save buttons and save on back and tapaway save, this method is unnecessary. it also causes a bug when trying to edit a cell at the very end of the tableview. the tableview tries to change the height of it's containing scrollview to make room for the keyboard, which then ends up end displaying some cells, which then causes editing to end.