better updates badging
badges should be shown only when an update is new to you
once you've seen it in the updates list view, the badge
should go away but there will remain a highlight indicating
that you haven't fully inspected the update yet. that
highlight is not guaranteed to stick around forever and
will get overwritten when you pull to refresh or the next
time you launch the app.