Update README.md to provide guidance to users.
There is nothing in the README.md to guide users on when and how to
submit issues for the app. This commit adds a link to the Google Group
for iNaturalist and a few simple tips on how to help make issues more
useful to the INaturalistIOS developers. At this point those tips are
focused on errors rather than, e.g., feature request issues, but could
easily be expanded.