use common_name not default_name in Taxon mapping
pick common name (what the app calls defaultName) from common_name
and not default_name, since it's in all the endpoints that we use,
while default_name isn't in all of the endpoints. bad naming in the
mapping here, not worth fixing and revving core data since we're
trying to move away from it.