allow editing of obs photo positions
this requires not uploading the entire observation
when all that's changed is just a field on one
of the children of the observation, since the obs
upload would clobber the local edits. we also send
updates children first, while posts are parent first.
we also can't globally dirty an observation after
you save the observation on the edit screen, since
you might have only changed one of the children.
deceptively large change for a tiny feature. for #255