be more careful in geocoder callback
make sure we're not capturing self strongly
also, make sure there isn't a race condition
between the user saving the obs and the geocoder callback
resulting in this standalone obs being added to realm
but then having the geocoder update trigger. so we
make sure to make this update in a realm transaction
just in case.