don't fetch me user right after account creation
it can take a little bit before the me user is available, but we
can make a me user directly out of the response from account creation
and the users entered input. if we fetch the me user too early
after account creation we get 422s back, and if we're depending on
that me user fetch to finish local signin (creating the local user
account, etc) then that's a problem. so be safer here.