replace INatWebController with ForgotPasswordController
under the hood, replace UIWebView (deprecated) with WKWebView
and dismiss the forgot password UI when we've POST'd and gotten
a 2xx or 3xx response from the server. should be more flexible
than hard coding the allowed urls in the web view, since that
can change if the server flow for forgot password changes (as
it has recently).

for #609